requirements.txt
django>=5.2,<6
python-decouple
pillow
djangorestframework
markdown
django-filter
drf-yasg
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
djoser
black
flake8
